Change RET action to jump to the `selectionRange` instead of the item's `Range`.
The current RET action jumps to the range of the `Item` returned from
`CallHierarchyXXXCall`. A more useful behaviour (and the one VS Code
uses) is to jump to the location of the first `fromRanges` (which is
where the function is actually called).

For e.g. if function A is called by B. The current incoming call
hierarchy for A will have B. Hitting RET on B will jump to
the *definition* of B, rather than to where A is called in function B.

It is possible for multiple `Range`s to be returned in the
`selectionRange` and for now we only jump to the first one. This is
similar to the behaviour of VS Code.